Give up long long, replace with short
Long long (and everything bigger than short, actually) breaks some
examples from http://esoteric.sange.fi/brainfuck/ (specifically, PI and
99 bottles). Sadly, I am not at all proficient in Brainfuck, so I can
not say if my interpreter or the code in question is at fault.

+++ brickfuck.c
@@ -73,8 +73,8 @@ void interpret_brainfuck(char *source){
 		return;
 	}
 
-	long long bf_tape[30000] = {0};
-	long long *bf_ptr = bf_tape;
+	short bf_tape[30000] = {0};
+	short *bf_ptr = bf_tape;
 	char *i = source;
 	char *stack[1024] = {0};
 	char **stack_ptr = stack;
